Refactor Walker to abstract file addition logic
Remove knowledge of valid and invalid extensions from Walker by
instead passing a filter function which encapsulates logic to determine
if a file should be added.

Create a FileInfo interface which implements a subset of os.FileInfo
to remove dependency on concrete types and allow for the re-use of the
existing os.FileInfo object passed to walkFn.

Add test coverage for the implemented filter function in Generator.